Removed cluster id. Introduced the mothership that will be main raft. Changed fsm to instance fsm so that we can also have a mothership fsm. Moved the add topic and remove topic calls to mothership.
RPC server is still broken. Added an instance selector that will select
the required broker for us.

Also started refactoring storage controller so that we can have
a topic config store to be used by mothership and some of the
same functionality in the storage controller as well since we
will need to still have a partition selector.